move detect and reconstruct into separat module
